refactor(blockchain): extract sync status tracker
